Extracted ServiceContainer create to a factory
