+++ Packages/com.daniil-ryzhkov.proceduraltoolkit/Services/DI/ServiceContainer.cs
@@ -24,18 +24,7 @@ namespace ProceduralToolkit.Services.DI
             this.singletonFactory = singletonFactory;
             this.transientFactory = transientFactory;
         }
-
-        public static IServiceContainer Create()
-        {
-            var list = new List<IService>();
-            var resolver = new ServiceResolver(list, () => new CycleProtector());
-            return new ServiceContainer(list,
-                                        resolver,
-                                        new ServiceInjector(resolver),
-                                        new SingletonServiceFactory(resolver),
-                                        new TransientServiceFactory(resolver));
-        }
-
+        
         public void AddSingleton<TImplementation>() where TImplementation : class
         {
             AddSingleton<TImplementation, TImplementation>();
