+++ Packages/com.daniil-ryzhkov.proceduraltoolkit/Services/DI/ServiceContainerFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2f0e52a3f848f4be09fcee9258e00773
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
